Refactor split big mod.rs to mod.rs and connection.rs
